chore: move top level modules into folders
