+++ lib/ref-resolver.js
