+++ lib/schema-validator.js
