+++ lib/serializer.js
