+++ lib/standalone.js
