+++ lib/validator.js
